+++ app/src/main/AndroidManifest.xml
@@ -8,22 +8,22 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".Main">
+        <activity android:name=".home">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".home" />
-        <activity android:name=".Menu" />
+        <activity android:name=".Main"
+            />
         <activity android:name=".Order" />
         <activity android:name=".menu_jc" />
         <activity android:name=".Menu_desrt" />
         <activity android:name=".OrderSignup" />
         <activity android:name=".OrderOrders" />
         <activity android:name=".OrderConfirmation" />
-        <activity android:name=".Main2" />
+
     </application>
 
 </manifest>
