--- app/src/main/java/com/fci/delite/Menu.java
@@ -1,13 +1,0 @@
-package com.fci.delite;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class Menu extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_menu);
-    }
-}
