+++ app/src/main/res/layout/activity_main.xml
@@ -7,6 +7,7 @@
     xmlns:design="http://schemas.android.com/apk/res-auto"
     tools:context="com.fci.delite.Main">
 
+
     <android.support.design.widget.BottomNavigationView
       android:id="@+id/bottomNavigationView"
         android:layout_width="match_parent"
