--- app/src/main/res/layout/activity_menu.xml
@@ -1,116 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.fci.delite.Menu">
-
-
-    <ImageButton
-            android:id="@+id/imageButton12"
-            android:layout_width="155dp"
-            android:layout_height="107dp"
-            app:srcCompat="@mipmap/ic_launcher"
-            tools:layout_editor_absoluteY="191dp"
-            android:layout_marginRight="1dp"
-            app:layout_constraintRight_toRightOf="parent"
-        android:layout_below="@+id/imageButton20"
-        android:layout_alignStart="@+id/imageButton19" />
-
-        <ImageButton
-            android:id="@+id/imageButton20"
-            android:layout_width="150dp"
-            android:layout_height="117dp"
-            app:srcCompat="@mipmap/ic_launcher"
-            tools:layout_editor_absoluteY="306dp"
-            android:layout_marginRight="8dp"
-            app:layout_constraintRight_toRightOf="parent"
-            android:layout_below="@+id/imageButton19"
-            android:layout_toEndOf="@+id/imageButton17" />
-
-        <ImageButton
-            android:id="@+id/imageButton19"
-            android:layout_width="158dp"
-            android:layout_height="115dp"
-            app:srcCompat="@mipmap/ic_launcher"
-            tools:layout_editor_absoluteY="76dp"
-            android:layout_marginRight="-2dp"
-            app:layout_constraintRight_toRightOf="parent"
-            android:layout_marginEnd="14dp"
-            android:layout_below="@+id/imageButton13"
-            android:layout_alignParentEnd="true" />
-
-        <ImageButton
-            android:id="@+id/imageButton18"
-            android:layout_width="221dp"
-            android:layout_height="192dp"
-            app:srcCompat="@mipmap/ic_launcher"
-            tools:layout_editor_absoluteX="0dp"
-            tools:layout_editor_absoluteY="243dp"
-            android:layout_alignParentTop="true"
-            android:layout_toStartOf="@+id/imageButton20"
-            android:layout_marginEnd="12dp" />
-
-        <ImageButton
-            android:id="@+id/imageButton17"
-            android:layout_width="210dp"
-            android:layout_height="178dp"
-            app:srcCompat="@mipmap/ic_launcher"
-            tools:layout_editor_absoluteX="0dp"
-            android:layout_above="@+id/linearLayout3"
-            android:layout_toStartOf="@+id/imageButton12"
-            android:layout_marginEnd="10dp"
-            android:layout_marginBottom="0dp"
-            app:layout_constraintBottom_toTopOf="@+id/imageButton18"
-            tools:layout_editor_absoluteY="65dp" />
-
-    <LinearLayout
-        android:id="@+id/linearLayout3"
-        android:layout_width="383dp"
-        android:layout_height="69dp"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentEnd="true"
-        android:alpha="0.5"
-        android:background="@android:color/holo_orange_light"
-        android:orientation="horizontal"
-        app:layout_constraintBottom_toBottomOf="parent"
-        tools:layout_editor_absoluteX="0dp"
-        tools:layout_editor_absoluteY="437dp">
-
-        <ImageButton
-            android:id="@+id/imgHome"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:srcCompat="@mipmap/ic_launcher"
-            tools:layout_editor_absoluteX="311dp"
-            tools:layout_editor_absoluteY="427dp" />
-
-        <ImageButton
-            android:id="@+id/imgMenu"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:srcCompat="@mipmap/ic_launcher"
-            tools:layout_editor_absoluteX="216dp"
-            tools:layout_editor_absoluteY="427dp" />
-
-        <ImageButton
-            android:id="@+id/imgOrder"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:srcCompat="@mipmap/ic_launcher"
-            tools:layout_editor_absoluteX="100dp"
-            tools:layout_editor_absoluteY="427dp" />
-
-        <ImageButton
-            android:id="@+id/imgMore"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:srcCompat="@mipmap/ic_launcher"
-            tools:layout_editor_absoluteX="0dp"
-            tools:layout_editor_absoluteY="427dp" />
-
-    </LinearLayout>
-
-
-</android.support.constraint.ConstraintLayout>
