+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
 <string name="app_name">Delite</string>
-<string name="title_activity_quick_nav">QuickNav</string>
+<string name="title_activity_quick_nav">D</string>
 
 <string name="navigation_drawer_open">Open navigation drawer</string>
 <string name="navigation_drawer_close">Close navigation drawer</string>
